task_manager/pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sustavov Maksim <maksim.sustavov@edgecenter.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
whitenoise = {extras = ["brotli"], version = "^6.8.2"}
gunicorn = "^23.0.0"
uvicorn = "^0.33.0"
django-bootstrap5 = "^24.3"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
